requirements.txt
Django==3.2.9
psycopg2-binary==2.8.6
flake8==3.9.0
flake8-django==1.1.1
pre-commit==2.11.1
python-dotenv~=0.17.0
black==21.7b0
djangorestframework~=3.12.0
